fix:create client component for about list to use custom lazy
