+++ src/app/[lang]/about/page.tsx
@@ -17,11 +17,12 @@ const DynamicFooter = dynamic(
 );
 const SideBar = dynamic(() => import("@/components/module/sidebar/SideBar"));
 const BreadCrumb = dynamic(() => import("@/components/shared/BreadCrumb"));
-const UserCard = dynamic(() => import("@/components/shared/UserCard"));
+const AboutList = dynamic(() => import("./components/list"));
 
 import useServerDarkMode from "src/hooks/use-server-dark-mode";
 import { staticMenuToShow as MenuStaticData } from "@/components/utils/constants";
 import Image from "next/image";
+import Head from "next/head";
 
 // SEO**
 export async function generateMetadata({ params }: any) {
@@ -98,171 +99,7 @@ export default async function AboutPage({ params }: any) {
   );
   const levelListArrayContent = await findByTabName(levelModals, "level-list");
 
-  function localFind1(_name: any) {
-    return citizenListArrayContent.find((item: any) => item.name == _name)
-      ?.translation;
-  }
-  function localFind2(_name: any) {
-    return levelListArrayContent.find((item: any) => item.name == _name)
-      ?.translation;
-  }
-
   const staticMenuToShow = MenuStaticData;
-
-  const staticUsers = [
-    {
-      id: 1,
-      name: "حسین قدیری",
-      profile_photo: "/profile/hossein-ghadiri.jpg",
-      code: "HM-2000001",
-      score: "",
-      levels: { current: { name: "بنیان گذار" } },
-    },
-    {
-      id: 2,
-      name: "امیر مدنی فر",
-      profile_photo: "",
-      code: "HM-2000002",
-      score: "",
-      levels: { current: { name: "بنیان گذار" } },
-    },
-    {
-      id: 3,
-      name: "عباس آجرلو",
-      profile_photo: "",
-      code: "HM-2000005",
-      score: "",
-      levels: { current: { name: "توسعه دهنده" } },
-    },
-    {
-      id: 4,
-      name: "مهدی غلام حسینی",
-      profile_photo: "",
-      code: "HM-2000008",
-      score: "",
-      levels: { current: { name: "توسعه دهنده" } },
-    },
-    {
-      id: 5,
-      name: "نازنین حشمتی",
-      profile_photo: "",
-      code: "",
-      score: "",
-      levels: { current: { name: "توسعه دهنده" } },
-    },
-    {
-      id: 6,
-      name: "امیر محسنی",
-      profile_photo: "",
-      code: "HM-2000475",
-      score: "",
-      levels: { current: { name: "توسعه دهنده" } },
-    },
-    {
-      id: 7,
-      name: "امین دهقان نژاد",
-      profile_photo: "",
-      code: "",
-      score: "",
-      levels: { current: { name: "توسعه دهنده" } },
-    },
-    {
-      id: 8,
-      name: "فاطمه نصیری",
-      profile_photo: "",
-      code: "",
-      score: "",
-      levels: { current: { name: "توسعه دهنده" } },
-    },
-    {
-      id: 9,
-      name: "بنیامین نوری",
-      profile_photo: "",
-      code: "HM-2000011",
-      score: "",
-      levels: { current: { name: "توسعه دهنده" } },
-    },
-    {
-      id: 10,
-      name: "مصطفی قدیری",
-      profile_photo: "",
-      code: "",
-      score: "",
-      levels: { current: { name: "توسعه دهنده" } },
-    },
-    {
-      id: 11,
-      name: "محمدجواد گرئی",
-      profile_photo: "",
-      code: "",
-      score: "",
-      levels: { current: { name: "توسعه دهنده" } },
-    },
-    {
-      id: 12,
-      name: "امیر حسین امینی",
-      profile_photo: "",
-      code: "HM-2000010",
-      score: "",
-      levels: { current: { name: "توسعه دهنده" } },
-    },
-    {
-      id: 13,
-      name: "آی تای ملکی",
-      profile_photo: "",
-      code: "",
-      score: "",
-      levels: { current: { name: "توسعه دهنده" } },
-    },
-    {
-      id: 14,
-      name: "یوسف خدری",
-      profile_photo: "",
-      code: "",
-      score: "",
-      levels: { current: { name: "توسعه دهنده" } },
-    },
-    {
-      id: 15,
-      name: "پرهام امین لو",
-      profile_photo: "",
-      code: "",
-      score: "",
-      levels: { current: { name: "توسعه دهنده" } },
-    },
-    {
-      id: 16,
-      name: "محمدرضا اصغری",
-      profile_photo: "",
-      code: "",
-      score: "",
-      levels: { current: { name: "توسعه دهنده" } },
-    },
-    {
-      id: 17,
-      name: "مرضیه ثاقب علیزاده",
-      profile_photo: "",
-      code: "HM-2000003",
-      score: "",
-      levels: { current: { name: "توسعه دهنده" } },
-    },
-    {
-      id: 18,
-      name: "سعید زاجکانی",
-      profile_photo: "",
-      code: "HM-2000009",
-      score: "",
-      levels: { current: { name: "توسعه دهنده" } },
-    },
-    {
-      id: 19,
-      name: "پارسا بهرامی",
-      profile_photo: "",
-      code: "HM-2000491",
-      score: "",
-      levels: { current: { name: "توسعه دهنده" } },
-    },
-  ];
 
   // add staticMenuToShow values to siblings tabsMenu values
   tabsMenu.forEach((tab: any) => {
@@ -302,6 +139,10 @@ export default async function AboutPage({ params }: any) {
 
   return (
     <>
+      <Head>
+        {/* Preload */}
+        <link rel="preload" href="/team.webp" as="image" />
+      </Head>
       {/* SCHEMA** */}
       <script
         type="application/ld+json"
@@ -469,20 +310,11 @@ export default async function AboutPage({ params }: any) {
                 </p>
               </div>
             </div>
-
-            <div className="flex flex-row flex-wrap justify-center md:justify-center w-full no-scrollbar overflow-y-auto py-[20px]">
-              {staticUsers.map((item: any, index: any) => (
-                <UserCard
-                  key={index}
-                  item={item}
-                  index={index}
-                  params={params}
-                  minWidth={`260px`}
-                  levelText={localFind2("developer")}
-                  buttonText={localFind1("citizen page")}
-                />
-              ))}
-            </div>
+            <AboutList
+              params={params}
+              citizenListArrayContent={citizenListArrayContent}
+              levelListArrayContent={levelListArrayContent}
+            />
           </section>
           <div className="xl:px-32 lg:px-32 md:px-5 sm:px-5 xs:px-1">
             <DynamicFooter footerTabs={footerTabs} />
