+++ src/app/[lang]/page.jsx
@@ -212,8 +212,8 @@ export default async function LangPage({params}) {
     <>
       <Head>
         {/* Preload poster image for video */}
-        <link rel="preload" href="/firstpage/replaced_pic.png" as="image" />
-        <link rel="preload" href="/firstpage/3d_rgb.irpsc.mp4" as="video" type="video/mp4" />
+        <link rel="preload" href="/firstpage/replaced_pic.webp" as="image" />
+        <link rel="preload" href="/firstpage/3d_rgb.irpsc.webm" as="video" type="video/mp4" />
       </Head>
       {/* SCHEMA** */}
       <script
